chore: modularized the code
