+++ backend/constants.py
@@ -1,0 +1,1 @@
+ACCESS_TOKEN_EXPIRE_MINUTES = 30
