--- backend/docker-compose.yml
@@ -1,27 +1,0 @@
-services:
-  citycare-backend:
-    build: .
-    ports:
-      - "8000:8000"
-    environment:
-      - DATABASE_URL=sqlite:///./citycare.db
-      - SECRET_KEY=your-super-secret-key-change-in-production
-    volumes:
-      - ./uploads:/app/uploads
-      - ./citycare.db:/app/citycare.db
-    restart: unless-stopped
-
-  # Optional: PostgreSQL database
-  # postgres:
-  #   image: postgres:15
-  #   environment:
-  #     POSTGRES_DB: citycare
-  #     POSTGRES_USER: citycare
-  #     POSTGRES_PASSWORD: citycare123
-  #   ports:
-  #     - "5432:5432"
-  #   volumes:
-  #     - postgres_data:/var/lib/postgresql/data
-
-# volumes:
-#   postgres_data:
